fix: startup crash, CI ARM64, audio portability
- Add srand(time(NULL)) and stage init in main.c (startup crash fix)
- Add road_get_segment length guard for empty stage
- Replace DEB822 CI sources with arch-tagged amd64/arm64 sources
- Replace sinf/floorf with sin/floor for macOS strict C11

+++ src/audio.c
@@ -5,9 +5,9 @@
 
 static float wave_generate(WaveType type, float phase) {
     switch (type) {
-        case WAVE_SINE: return sinf(phase * 2.0f * PI);
-        case WAVE_SAWTOOTH: return 2.0f * (phase - floorf(phase + 0.5f));
-        case WAVE_SQUARE: return (sinf(phase * 2.0f * PI) >= 0) ? 1.0f : -1.0f;
+        case WAVE_SINE: return sin(phase * 2.0f * PI);
+        case WAVE_SAWTOOTH: return 2.0f * (phase - floor(phase + 0.5f));
+        case WAVE_SQUARE: return (sin(phase * 2.0f * PI) >= 0) ? 1.0f : -1.0f;
         case WAVE_NOISE: return ((float)rand() / RAND_MAX) * 2.0f - 1.0f;
         default: return 0.0f;
     }
